requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==25.1.0
certifi==2024.12.14
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
yarl==1.18.3
